+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Now if no `cliName` passed, it sets `unnamed-cli`
 - Updated return value of args from string to object
 - Added an async util `navigateSubcommands`
+- Made `treeGraph` async
 
 ## v2.1.0 2024-07-18